Actually activate inventory caching (#97)
The inventory plugin advertises support for caching, but the
implementation was missing the required pieces to make it work
properly.

This commit split the inventory construction into two stages. The
first stage is responsible for fetching remote data (interacting with
the DigitalOcean's API). The second stage extracts the relevant data
from the API responses and populated the inventory.

Caching support then came almost for free. All we had to do is only
fetch data if:

 1. User does not want to use cache at all.
 2. User does use cache and wants data refreshed.
 3. User does use cache but the cache is empty.

There is some additional logic that makes sure we write the data to
cache if data refresh is requested, but nothing too complex.