Allow deleting droplet by id
For droplet deletion, #184 required `name` and `unique_name` and
verified `unique_name` is true, to the exclusion of `id`.

Remove requirement from `required_if` as it cannot enforce the logic
`id or (name and unique_name)` and modify the verification of
`unique_name` to include `id` (and don't worry about `name` because
either it or `id` are already required elsewhere).

Fixes #260.