droplet: Allow creating droplet without droplet agent
Allow passing `with_droplet_agent` boolean argument to droplet create
API request to explicitly install or not the droplet agent which
facilitates connecting to the droplet using the web console in the
control panel.  The API default is to install the droplet agent.  This
option allows a user to request not installing the droplet agent.

https://docs.digitalocean.com/reference/api/api-reference/#operation/droplets_create